Availability/quantity feature & edits
- Added the Integer variable quantity in Class Book
- Updated the checkout feature so that whenever a checkout happens, all
the books that were in the cart get their quantities reduced by 1
- Updated all the necessary Html templates to show the quantities and to
prevent buying books with 0 quantity, and added an Availability boolean
attribute which determines whether a book can be placed in cart or not
based on the quantity.
- Added necessary links between the Html templates
- Added a default error Html template
- Changed the repositories from extending PagingAndSortingRepository to
extending JpaRepository which extends the former and allows for usage of
the saveAndFlush() method.
- Updated the test classes, and the integration test to ensure that the
books' quantities are reduced after a checkout happens.
- updated the CommandLineRunner args.